fix incorrect use of "inject"
"inject", as opposed to "spawn" has a clearly defined meaning per
doc/TCL.md. The usage of the term "inject" was not consistently used,
and in several placed was used to refer to spawning. This has been
rectified.

Namely, this renames the existing "inject" instrument to "spawn". A new
inject instrument is added to align with the proper meaning. Also, the
"inject" command is now named "spawn", although an alias named "inject"
is preserved for convenience.

Additionally, many performance counters are added both globally and to
notes.