add nocviz_graph_foreach_link
This required adding a vector to store all links in the graph in, for
easy iteration. The only alternative would have been to iterate over
each node and deduplicate links that have already been seen, which would
have been expensive and complex.

Tests have been added to ensure this does not introduce a memory leak.